Fixed incorrect symmetry operator count when finding FZ orientations for Cubic
Took the opportunity to clean up the symmetry operators and convert them from "C"
style 2D arrays into std::vector of QuatD or OrientationD types. This cleans up the
API in some of the called methods where we don't have to send in a count of the
symmetry operators possibly introducing errors (like what was fixed).

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>